server/requirements.txt
Flask==2.3.2
requests==2.31.0
google-generativeai==0.2.0

